ipvlan: implement MTU handling
Currently, the ipvlan driver has some limitations with respect
to the handling of MTU:
- there is no way set MTU greater than 1500 via MTU change
path (because max_mtu is not redefined in the driver)
- the driver does not handle IFLA_MTU, so it's not possible
to specify initial MTU at link creation time
- the increase of MTU on the parent device automatically
increases the MTU on the ipvlan device via NETDEV_CHANGEMTU
event even if the MTU on the ipvlan device was previously
intentionally decreased (I cannot think about the use-case
where this behavior is desirable and most other drivers
explicitly prohibit auto-increase of the MTU on devices
as a result of parent MTU increase)

This commit implements MTU handling for the ipvlan driver:
- min_mtu and max_mtu are explicitly redefined. max_mtu is
initialized to ETH_MAX_MTU. This allows to explicitly set
MTU greater than 1500
- added handling of IFLA_MTU, so it's now possible to
define the MTU of ipvlan at creation time, e.g. this will
work now:
ip link add mtu 1400 link enp0s3 ipvl0 type ipvlan
- implemented the ndo_change_mtu callback for the ipvlan driver.
This callback does not allow you to set MTU greater than
MTU of the parent device
- Updated NETDEV_CHANGEMTU handler so that it no longer
automatically increases the MTU as a result of the increase of
the MTU on the parent device

I decided to bump the version of the driver as this change
removes the "feature" of auto increase of the MTU.
Technically, it's a backward incompatible change, so if that's
a concern, then the new logic may be made conditional so that users
need to explicitly enable it via the driver option or sysfs parameter.
Otherwise, user-space code that depends on this feature may just
parse "ethtool -i" output to figure out the version of the driver
and do the right thing.

P.S.
I've tested all possible combinations of ipvlan modes in the default
and custom namespaces with respect to a change of the MTU on:
* the ipvlan device itself at device creation time
* the ipvlan device itself via a normal path, i.e. dev_set_mtu()
* the parent device of existing ipvlan devices